feat: update dots_indicator to v4.0.1 and add loading state for localization
- Updated the `dots_indicator` package to version 4.0.1 to leverage the latest features and improvements.
- Added `AppLocalizationLoading` state to handle loading during language changes, improving user experience.
- Fixed type casting for `DotsIndicator` position to double to ensure compatibility with the updated package.

+++ lib/features/shop/presentation/widgets/shop_offers_items.dart
@@ -62,7 +62,7 @@ class _ShopOffersItemsState extends State<ShopOffersItems> {
           valueListenable: _currentOffer,
           builder: (context, value, child) => DotsIndicator(
             dotsCount: 4,
-            position: value,
+            position: value.toDouble(),
             decorator: const DotsDecorator(
               color: Color(0xffEAE8E8),
               size: Size.fromRadius(6),
